+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/FLTGoogleMapTileOverlayController.h
@@ -5,6 +5,8 @@
 #import <Flutter/Flutter.h>
 #import <GoogleMaps/GoogleMaps.h>
 
+#import "messages.g.h"
+
 NS_ASSUME_NONNULL_BEGIN
 
 @interface FLTGoogleMapTileOverlayController : NSObject
@@ -20,17 +22,18 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface FLTTileProviderController : GMSTileLayer
 @property(copy, nonatomic, readonly) NSString *tileOverlayIdentifier;
-- (instancetype)init:(FlutterMethodChannel *)methodChannel
-    withTileOverlayIdentifier:(NSString *)identifier;
+- (instancetype)initWithTileOverlayIdentifier:(NSString *)identifier
+                              callbackHandler:(FGMMapsCallbackApi *)callbackHandler;
 @end
 
 @interface FLTTileOverlaysController : NSObject
-- (instancetype)init:(FlutterMethodChannel *)methodChannel
-             mapView:(GMSMapView *)mapView
-           registrar:(NSObject<FlutterPluginRegistrar> *)registrar;
-- (void)addTileOverlays:(NSArray *)tileOverlaysToAdd;
-- (void)changeTileOverlays:(NSArray *)tileOverlaysToChange;
-- (void)removeTileOverlayWithIdentifiers:(NSArray *)identifiers;
+- (instancetype)initWithMapView:(GMSMapView *)mapView
+                callbackHandler:(FGMMapsCallbackApi *)callbackHandler
+                      registrar:(NSObject<FlutterPluginRegistrar> *)registrar;
+- (void)addJSONTileOverlays:(NSArray<NSDictionary<NSString *, id> *> *)tileOverlaysToAdd;
+- (void)addTileOverlays:(NSArray<FGMPlatformTileOverlay *> *)tileOverlaysToAdd;
+- (void)changeTileOverlays:(NSArray<FGMPlatformTileOverlay *> *)tileOverlaysToChange;
+- (void)removeTileOverlayWithIdentifiers:(NSArray<NSString *> *)identifiers;
 - (void)clearTileCacheWithIdentifier:(NSString *)identifier;
 - (nullable FLTGoogleMapTileOverlayController *)tileOverlayWithIdentifier:(NSString *)identifier;
 @end