+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.10.0
+
+* Converts Obj-C->Dart calls to Pigeon.
+
+## 2.9.0
+
+* Converts additional platform calls to Pigeon.
+
 ## 2.8.2
 
 * Converts inspector interface platform calls to Pigeon.